update property table
