+++ src/views/dashboard/properties/PropTable.js
@@ -8,7 +8,9 @@ import {
   TableHead,
   TableRow
 } from '@material-ui/core';
-import DeleteIcon from '@material-ui/icons/Delete';
+import EditIcon from '@material-ui/icons/Edit';
+import { navigate } from '@reach/router';
+import PropertyDetailsModal from '../../../components/Properties/PropertyDetailsModal';
 
 const useStyles = makeStyles({
   table: {
@@ -23,29 +25,75 @@ export default function PropTable() {
 
   const propertyList = useSelector(state => state.propReducer.properties);
 
+  // open/close state for modal
+  const [openDetails, setOpenDetails] = useState(false);
+
+  // store individual property from map function in local state
+  const [currentProperty, setCurrentProperty] = useState({});
+
+  // handle open/close, takes in the individual property
+  // passed up from property card to gain access from the modal
+  const handleOpen = property => {
+    // sets the currentProperty state to the property object passed in
+    setCurrentProperty(property);
+
+    // this sets the open state for the PropertyDetailsModal
+    setOpenDetails(true);
+  };
+
+  const handleClose = () => {
+    // sets currentProperty to empty obj when modal is closed
+    setCurrentProperty({});
+
+    // sets the open state to false to close PropertyDetailsModal
+    setOpenDetails(false);
+  };
+
   return (
-    <Table className={classes.table} aria-label="Property Table">
-      <TableHead>
-        <TableRow>
-          <TableCell>Properties</TableCell>
-          <TableCell align="right">Address</TableCell>
-          <TableCell align="right">Status</TableCell>
-        </TableRow>
-      </TableHead>
-      <TableBody>
-        {propertyList.map(({ id, name, street, status }) => (
-          <TableRow key={id}>
-            <TableCell component="th" scope="row">
-              {name}
-            </TableCell>
-            <TableCell align="right">{street}</TableCell>
-            <TableCell align="right">{status}</TableCell>
-            <TableCell align="right">
-              <DeleteIcon />
-            </TableCell>
+    <>
+      <Table className={classes.table} aria-label="Property Table">
+        <TableHead>
+          <TableRow>
+            <TableCell>Id</TableCell>
+            <TableCell align="right">Name</TableCell>
+            <TableCell align="right">Street</TableCell>
+            <TableCell align="right">State</TableCell>
+            <TableCell align="right">Status</TableCell>
+            <TableCell align="right">Edit</TableCell>
           </TableRow>
-        ))}
-      </TableBody>
-    </Table>
+        </TableHead>
+        <TableBody>
+          {propertyList.map(prop => (
+            <TableRow
+              key={prop.id}
+              onClick={() => {
+                navigate(`/dashboard/properties/${prop.id}`);
+                handleOpen(prop);
+              }}
+            >
+              <TableCell>{prop.id}</TableCell>
+              <TableCell align="right">{prop.name}</TableCell>
+              <TableCell align="right">{prop.street}</TableCell>
+              <TableCell align="right">{prop.state}</TableCell>
+              <TableCell align="right">{prop.status}</TableCell>
+              <TableCell align="right">
+                <EditIcon
+                  style={{ cursor: 'pointer' }}
+                  onClick={e => {
+                    e.stopPropagation();
+                    navigate(`/dashboard/properties/edit/${prop.id}`);
+                  }}
+                />
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+      <PropertyDetailsModal
+        property={currentProperty}
+        open={openDetails}
+        close={handleClose}
+      />
+    </>
   );
 }
